More code clean up
* Import code clean up

* Correct PEP8 E401 errors

E401 multiple imports on one line

* Line continuation code clean up

* Correct PEP8 E502 errors

E502 the backslash is redundant between brackets

* Code structure clean up

* Correct PEP8 E7xx errors

E701 multiple statements on one line (colon)
E702 multiple statements on one line (semicolon)
E703 statement ends with a semicolon
E711 comparison to None should be 'if cond is None:'
E711 comparison to None should be 'if cond is not None:'
E712 comparison to False should be 'if cond is False:' or 'if not cond:'
E712 comparison to False should be 'if cond is not False:' or 'if cond:'
E712 comparison to True should be 'if cond is True:' or 'if cond:'
E712 comparison to True should be 'if cond is not True:' or 'if not cond:'
E713 test for membership should be 'not in'
E714 test for object identity should be 'is not'
E722 do not use bare 'except'
E731 do not assign a lambda expression, use a def
E741 ambiguous variable name 'l'

NOTE: Some manual clean up was required after the E731 corrections.  These would probably be corrected in a more complete / comprehensive autopep8 run.