Major PEP8 clean up
This clean up is the result of running the following two commands across the repository:
- python /media/hdd/dev/autopep8-1.5.5/autopep8.py -a -a -a -i -r --ignore E402,E501 <repository_dir>
- find <repository_dir> -type f -name "*.py" | xargs sed -i -r -e ":a;s/^(\t*)\ {4}/\1\t/;ta" -e "s/^([^#].*?\S)\s*#/\1  #/" -e "s/^(\s*)#([^!#\ /])/\1#\ \2/g" -e "s/#\t/# \t/g" -e "s/\"  #(.*?\")/\"#\1/g"

I then had to manually correct Screens/VirtualKeyBoard.py line 54 and Plugins/SystemPlugins/SoftwareManager/plugin.py line 230 to restore the missing two spaces before the comment "#".

The indentation around multiline statements is not optimal but I am hoping that this is a good start to the code clean up to be continued with the flake8 lint tool.